+++ packages/gits_http/lib/src/cache_strategy/storage/cache_storage.dart
@@ -1,15 +1,27 @@
 import 'package:hive/hive.dart';
+import 'package:path_provider/path_provider.dart';
 
 import 'storage.dart';
 
 class CacheStorage implements Storage {
   static const _hiveBoxName = "gits_http_cache_storage";
 
+  bool _hasInit = false;
+
   CacheStorage();
+
+  Future<dynamic> openBox() async {
+    if (!_hasInit) {
+      final dir = await getApplicationDocumentsDirectory();
+      Hive.init(dir.path);
+      _hasInit = true;
+    }
+    return Hive.openBox(_hiveBoxName);
+  }
 
   @override
   Future<void> clear({String? prefix}) async {
-    final box = await Hive.openBox(_hiveBoxName);
+    final box = await openBox();
     if (prefix == null) {
       await box.clear();
     } else {
@@ -23,25 +35,25 @@ class CacheStorage implements Storage {
 
   @override
   Future<void> delete(String key) async {
-    final box = await Hive.openBox(_hiveBoxName);
+    final box = await openBox();
     return box.delete(key);
   }
 
   @override
   Future<String?> read(String key) async {
-    final box = await Hive.openBox(_hiveBoxName);
+    final box = await openBox();
     return box.get(key);
   }
 
   @override
   Future<void> write(String key, String value) async {
-    final box = await Hive.openBox(_hiveBoxName);
+    final box = await openBox();
     return box.put(key, value);
   }
 
   @override
   Future<int> count({String? prefix}) async {
-    final box = await Hive.openBox(_hiveBoxName);
+    final box = await openBox();
     if (prefix == null) {
       return box.length;
     } else {